move-only: move definitions to clientfeeds.{cpp,h} to deal with circulars
