+++ src/Makefile.qt.include
@@ -244,6 +244,8 @@ BITCOIN_QT_BASE_CPP = \
   qt/bantablemodel.cpp \
   qt/bitcoin.cpp \
   qt/bitcoinaddressvalidator.cpp \
+  qt/masternodemodel.cpp \
+  qt/proposalmodel.cpp \
   qt/bitcoinamountfield.cpp \
   qt/bitcoingui.cpp \
   qt/bitcoinunits.cpp \
@@ -281,14 +283,12 @@ BITCOIN_QT_WALLET_CPP = \
   qt/descriptiondialog.cpp \
   qt/editaddressdialog.cpp \
   qt/masternodelist.cpp \
-  qt/masternodemodel.cpp \
   qt/mnemonicverificationdialog.cpp \
   qt/openuridialog.cpp \
   qt/overviewpage.cpp \
   qt/paymentserver.cpp \
   qt/proposalcreate.cpp \
   qt/proposallist.cpp \
-  qt/proposalmodel.cpp \
   qt/proposalresume.cpp \
   qt/psbtoperationsdialog.cpp \
   qt/qrdialog.cpp \
